Implements comment deleting.

+++ project/libs/blog/blog-comment/src/blog-comment-module/blog-comment.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 
 import { BlogCommentRepository } from './blog-comment.repository';
 import { BlogCommentEntity } from './blog-comment.entity';
@@ -31,25 +35,21 @@ export class BlogCommentService {
     return newComment;
   }
 
-  public async deleteComment(id: string): Promise<void> {
+  public async deleteComment(id: string, authorId: string): Promise<void> {
+    const existsComment = await this.blogCommentRepository.findById(id);
+
+    if (!existsComment) {
+      throw new NotFoundException(`Comment with id ${id} was not found.`);
+    }
+
+    if (existsComment.authorId !== authorId) {
+      throw new ForbiddenException('Users can only delete their own comments.');
+    }
+
     try {
       await this.blogCommentRepository.deleteById(id);
     } catch {
       throw new NotFoundException(`Comment with ID ${id} not found`);
     }
   }
-
-  // public async updateComment(
-  //   id: string,
-  //   dto: UpdateCommentDto
-  // ): Promise<BlogCommentEntity> {
-  //   const blogCommentEntity = new BlogCommentEntity(dto);
-
-  //   try {
-  //     await this.blogCommentRepository.update(blogCommentEntity);
-  //     return blogCommentEntity;
-  //   } catch {
-  //     throw new NotFoundException(`Comment with ID ${id} not found`);
-  //   }
-  // }
 }
